requirements.txt
pyttsx3>=2.90
google-generativeai>=0.8.6
